RAG/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
chromadb>=0.5.15
